# credit-service/src/main/resources/openapi/credit-application.yaml
openapi: 3.0.3
info:
  title: Credit Application API
  description: API для управления заявками на кредит
  version: 1.0.0
servers:
  - url: http://localhost:8081/api/v1
    description: Локальный сервер

paths:
  /credit-applications:
    post:
      summary: Создать новую заявку на кредит
      description: Создает новую заявку на кредит с указанными данными.
      operationId: createCreditApplication
      requestBody:
        description: Данные новой заявки на кредит
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditApplicationRequest'
      responses:
        '201':
          description: Заявка успешно создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditApplicationResponse'
        '400':
          description: Неверный запрос

    get:
      summary: Получить список заявок на кредит
      description: Возвращает список всех заявок на кредит.
      operationId: getAllCreditApplications
      responses:
        '200':
          description: Список заявок на кредит
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CreditApplicationResponse'

  /credit-applications/{id}/status:
    put:
      summary: Обновить статус заявки на кредит
      description: Обновляет статус существующей заявки.
      operationId: updateCreditApplicationStatus
      parameters:
        - name: id
          in: path
          required: true
          description: Уникальный идентификатор заявки
          schema:
            type: string
            format: uuid
      requestBody:
        description: Новый статус заявки
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  example: "APPROVED"
              required:
                - status
      responses:
        '200':
          description: Статус заявки успешно обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditApplicationResponse'
        '404':
          description: Заявка не найдена

components:
  schemas:
    CreditApplicationRequest:
      type: object
      properties:
        amount:
          type: number
          format: double
          example: 50000
        termInMonths:
          type: integer
          example: 12
      required:
        - amount
        - termInMonths

    CreditApplicationResponse:
      type: object
      properties:
        id:
          type: string
          example: "123e4567-e89b-12d3-a456-426614174000"
        amount:
          type: number
          format: double
          example: 50000
        termInMonths:
          type: integer
          example: 12
        status:
          type: string
          example: "PENDING"
        createdAt:
          type: string
          format: date-time
          example: "2024-06-16T14:32:00Z"